Refactor operator overloading code
