+++ source/framework/api/cxx/include/lue/framework/api/cxx.hpp
@@ -2,5 +2,4 @@
 #include "lue/framework/api/cxx/io.hpp"
 #include "lue/framework/api/cxx/local_operation.hpp"
 #include "lue/framework/api/cxx/miscellaneous_operation.hpp"
-#include "lue/framework/api/cxx/operator.hpp"
 #include "lue/framework/api/cxx/routing_operation.hpp"
