+++ source/framework/api/cxx/include/lue/framework/api/cxx/local/logical_not.hpp
@@ -7,4 +7,10 @@ namespace lue::api {
 
     auto LUE_FAPI_EXPORT logical_not(Field const& field) -> Field;
 
+
+    inline auto operator!(lue::api::Field const& argument)
+    {
+        return lue::api::logical_not(argument);
+    }
+
 }  // namespace lue::api
