+++ source/framework/api/cxx/include/lue/framework/api/cxx/local/modulus.hpp
@@ -7,4 +7,10 @@ namespace lue::api {
 
     auto LUE_FAPI_EXPORT modulus(Field const& field1, Field const& field2) -> Field;
 
+
+    inline auto operator%(lue::api::Field const& lhs, lue::api::Field const& rhs)
+    {
+        return lue::api::modulus(lhs, rhs);
+    }
+
 }  // namespace lue::api
