+++ source/framework/api/cxx/test/local/add_test.cpp
@@ -1,5 +1,5 @@
 #define BOOST_TEST_MODULE lue framework api cxx local add
-#include "lue/framework/api/cxx/operator.hpp"
+#include "lue/framework/api/cxx/local/add.hpp"
 #include "lue/framework/test/hpx_unit_test.hpp"
 #include "lue/framework.hpp"
