+++ source/framework/api/cxx/test/local/greater_than_equal_to_test.cpp
@@ -1,5 +1,5 @@
 #define BOOST_TEST_MODULE lue framework api cxx local greater_than_equal_to
-#include "lue/framework/api/cxx/operator.hpp"
+#include "lue/framework/api/cxx/local/greater_than_equal_to.hpp"
 #include "lue/framework/test/hpx_unit_test.hpp"
 #include "lue/framework.hpp"
