+++ source/framework/api/cxx/test/local/logical_not_test.cpp
@@ -1,5 +1,5 @@
 #define BOOST_TEST_MODULE lue framework api cxx local logical_not
-#include "lue/framework/api/cxx/operator.hpp"
+#include "lue/framework/api/cxx/local/logical_not.hpp"
 #include "lue/framework/test/hpx_unit_test.hpp"
 #include "lue/framework.hpp"
